% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_build_density.R
\name{export_build_density}
\alias{export_build_density}
\title{Export classified building types to a GPKG file}
\usage{
export_build_density(data, filename)
}
\arguments{
\item{data}{#resultant file from the "calculate_building_density" function}

\item{filename}{#name of the output GPKG file}
}
\description{
Export classified building types to a GPKG file
}
