% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_classified_gpkg.R
\name{export_classified_gpkg}
\alias{export_classified_gpkg}
\title{Export classified building types to a GPKG file}
\usage{
export_classified_gpkg(data, filename)
}
\arguments{
\item{data}{#resultant file from the "classify_osm_buildings" function}

\item{filename}{#name of the output GPKG file}
}
\description{
Export classified building types to a GPKG file
}
