% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_build_density.R
\name{plot_building_density}
\alias{plot_building_density}
\title{Title}
\usage{
plot_building_density(build_density_GDF, city_name = "", density_column)
}
\arguments{
\item{city_name}{#name of the city}

\item{gdf}{#resultant file from the "calculate_building_density" function}

\item{attribute}{#name of the attribute to be plotted, build_density in this case}
}
\value{
# Choropleth map of building density
}
\description{
Title
}
\examples{
#' city_name <- "Wuerzburg"
#' density_column <- "build_density"
#gdf <- calculate_building_density(city_name, admin_level = 8, bbox = NULL)


}
